zzz_analytics-audit-engine/requirements.txt
# Core Dependencies
python-dotenv==1.0.0

# Web Crawling & Scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
aiohttp==3.9.1

# JavaScript Parsing
js2py==0.74
esprima==4.0.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# CLI
click==8.1.7
rich==13.7.0
typer==0.9.0

# Reporting
jinja2==3.1.2
weasyprint==60.1
markdown==3.5.1
plotly==5.18.0

# Dashboard
streamlit==1.29.0
dash==2.14.2
dash-bootstrap-components==1.5.0

# API
flask==3.0.0
flask-cors==4.0.0
flask-restful==0.3.10

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1
validators==0.22.0
tld==0.13
urllib3==2.1.0

# Privacy/Consent Detection
advertools==0.14.2

# Performance Monitoring
psutil==5.9.6

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Logging
structlog==23.2.0
colorlog==6.8.0
